Adds support for upgrading artifacts
We're effectively just adding multiple copies of a gift to an artifact, and only displaying the latest one.

Changes:
character_tags.py: When displaying artfacts on the player's inventory, only display the latest copy of a gift.
Comments: Does this by assuming that the highest unique ID on a crafting event is newest, as it doesn't have a publish date.

hgapp/crafting/models.py: Actual logic for calculating server-side cost of upgrades - does this by checking if an existing copy of the power exists already on the artfact, if it does, gets the delta, if it doesn't uses the normal cost.  When refunding, only unlinks the main copy of the power if there's no copies of it on the power.

crafting.js: User side code for displaying the powers for upgrade

hgapp/crafting/views.py: server-side calculation of powers that can be upgraded

hgapp/powers/models.py: helper function for getting the delta between two powers.